fix(cli-sub-agent): persist preference-aware fallback chains
Refs #1749.

Address pre-PR review round 5 findings (module token-budget split + preference-aware fallback chain).

Split run_cmd_tool_selection tests into a sibling module to keep the parent under the token budget, and thread tier preference order into fallback-chain persistence for review and debate so result.toml reflects the actual execution order.
